Spotify-Clone
Step into the world of Spotify-Clone, a pixel-perfect Spotify clone crafted solely with HTML and CSS. Experience a responsive, user-friendly interface that showcases the power of clean, modern web design. Whether you're exploring new tracks or jamming to classics, Spotify-Clone is a visual delight that hits all the right notes!